requirements.txt
python-dotenv~=0.18.0
celery~=4.4.7